refactor(assetlock): replace persisted status string union with AssetLockFundingStatus enum, move funding state type to types/
