+++ src/main/src/api/wallet/startAssetLockFunding.ts
@@ -1,5 +1,6 @@
 import { IpcMainInvokeEvent } from 'electron/utility'
-import { AssetLockService, AssetLockFundingState } from '../../services/AssetLockService'
+import { AssetLockService } from '../../services/AssetLockService'
+import { AssetLockFundingState } from '../../types/AssetLockFunding'
 import { AssetLockFundingKind } from '../../database/AssetLockDAO'
 
 export class StartAssetLockFundingHandler {
