+++ src/main/src/enums/AssetLockFundingStatus.ts
@@ -1,0 +1,7 @@
+export enum AssetLockFundingStatus {
+  L1Broadcast = 'l1_broadcast',
+  ChainLocked = 'chainlocked',
+  StBroadcast = 'st_broadcast',
+  Done = 'done',
+  Error = 'error',
+}
